+++ src/snowflake/snowpark/dataframe_writer.py
@@ -2,7 +2,7 @@
 # Copyright (c) 2012-2022 Snowflake Computing Inc. All rights reserved.
 #
 
-from typing import Dict, Iterable, List, Optional, Union
+from typing import Any, Dict, Iterable, List, Optional, Union
 
 import snowflake.snowpark  # for forward references of type hints
 from snowflake.snowpark._internal.analyzer.snowflake_plan_node import (
@@ -68,6 +68,7 @@ class DataFrameWriter:
         *,
         mode: Optional[str] = None,
         create_temp_table: bool = False,
+        _statement_params: Optional[Dict[str, Any]] = None,
     ) -> None:
         """Writes the data to the specified table in a Snowflake database.
 
@@ -86,6 +87,7 @@ class DataFrameWriter:
                 "ignore": Ignore this operation if data already exists.
 
             create_temp_table: The to-be-created table will be temporary if this is set to ``True``.
+            _statement_params: Extra information that should be sent to Snowflake with query.
 
         Examples::
 
@@ -112,7 +114,7 @@ class DataFrameWriter:
         )
         session = self._dataframe._session
         snowflake_plan = session._analyzer.resolve(create_table_logic_plan)
-        session._conn.execute(snowflake_plan)
+        session._conn.execute(snowflake_plan, _statement_params=_statement_params)
 
     def copy_into_location(
         self,
@@ -123,6 +125,7 @@ class DataFrameWriter:
         file_format_type: Optional[str] = None,
         format_type_options: Optional[Dict[str, str]] = None,
         header: bool = False,
+        _statement_params: Optional[Dict[str, Any]] = None,
         **copy_options: Optional[str],
     ) -> List[Row]:
         """Executes a `COPY INTO <location> <https://docs.snowflake.com/en/sql-reference/sql/copy-into-location.html>`__ to unload data from a ``DataFrame`` into one or more files in a stage or external stage.
@@ -134,6 +137,7 @@ class DataFrameWriter:
             file_format_type: Specifies the type of files unloaded from the table. If a format type is specified, additional format-specific options can be specified in ``format_type_options``.
             format_type_options: Depending on the ``file_format_type`` specified, you can include more format specific options. Use the options documented in the `Format Type Options <https://docs.snowflake.com/en/sql-reference/sql/copy-into-location.html#format-type-options-formattypeoptions>`__.
             header: Specifies whether to include the table column headings in the output files.
+            _statement_params: Extra information that should be sent to Snowflake with query.
             copy_options: The kwargs that are used to specify the copy options. Use the options documented in the `Copy Options <https://docs.snowflake.com/en/sql-reference/sql/copy-into-location.html#copy-options-copyoptions>`__.
 
         Returns:
@@ -182,6 +186,6 @@ class DataFrameWriter:
                 copy_options=copy_options,
                 header=header,
             )
-        )._internal_collect_with_tag()
+        )._internal_collect_with_tag(_statement_params=_statement_params)
 
     saveAsTable = save_as_table